+++ keebgen/switch_socket.py
@@ -1,15 +1,14 @@
-from better_abc import ABC, abstractmethod
 import solid as sl
 from pathlib import Path
 
-from keebgen.geometry_base import Solid, CuboidAnchorCollection
+from keebgen.geometry_base import Part, CuboidAnchorCollection
 
 # when adding new sockets, the top of the socket should be coplanar with the X,Y plane
 # when the switch is installed, the keycap mounting feature should align with the Z axis
 
-class CherryMXSocket(Solid):
+class CherryMXSocket(Part):
     def __init__(self, config, u=1):
-        super(CherryMXSocket, self).__init__()
+        super().__init__()
         # determines how much flat space to reserve around the switch
         # prevents interference between keycap and other geometry
         width = config.getfloat('overall_width') + (u-1) * 19.0
@@ -39,9 +38,9 @@ class CherryMXSocket(Solid):
             side_nub = sl.cylinder(side_nub_radius, side_nub_width, segments=20, center=True)
             side_nub = sl.rotate(90, [1, 0, 0])(side_nub)
             side_nub = sl.translate([switch_width/2, 0, side_nub_radius-thickness])(side_nub)
-            nub_cube_len = (width-switch_width)/2
+            nub_cube_len = (width - switch_width)/2
             nub_cube = sl.cube([nub_cube_len, side_nub_width, thickness], center=True)
-            nub_cube = sl.translate([(width-nub_cube_len)/2, 0, -thickness/2])(nub_cube)
+            nub_cube = sl.translate([(width - nub_cube_len)/2, 0, -thickness/2])(nub_cube)
             side_nub = sl.hull()(side_nub, nub_cube)
 
             socket += side_nub + sl.rotate([0, 0, 180])(side_nub)
